[refactor] follow repository와 user 관련 repository 분리 (#80)
